Find the proper function to apply to each statistical and column block pairing and use the metadata associated with each column block for the function's arguments

Find the proper function to apply to each statistical and column block pairing and use the metadata associated with each column block for the function's arguments

CDISC_methods(agg, column, week, group, data)

Arguments

agg
the statistic to apply given the block name

column
the column to apply that statistic too, and class of the column dictated by the data frame it came from

week
the week if needed for calculation

group
whether to perform a group_by and if so by which column

data
the dataset to perform all functions on

Value

the table corresponding to the proper function to perform given the supplied column. This is used within a map to apply to all blocks inside the table generator module.

See also