Plot Individual Explorer Visits by Param

Gather various shiny inputs and return either a ggplot or plotly object. This graph's purpose is to plot AVAL against a visit variable for any PARAMCD, establishing one part of patient narrative found on the Individual Explorer Tab. The graph allows many features to be toggled to enhance the data presented in the plot, including plotting horizontal lines for screening & baseline measurements (when available) and overlaying pertinent OCCDs events (certain milestones, adverse events, and/or con meds) that have occurred during the patients journey.

fnIndvExplVisits(
  watermark = FALSE,
  graph_output = "plotly",
  bds_data,
  usubjid,
  input_plot_hor,
  input_visit_var,
  input_plot_param,
  input_plot_adam,
  input_overlay_events,
  vline_dat,
  vv_dy_name
)

Arguments

watermark
If TRUE, then include a watermark on the output plot

graph_output
A character string specifying either "plotly" or "ggplot2"

bds_data
A character string containg the name of the BDS Dataset

usubjid
A character string containing Patient number in the form of USUBJID standards

input_plot_hor
A character string containing variable name to have horizontal line plotted

input_visit_var
A character string containing the visit variable name

input_plot_param
A character string containing the PARAMCD name / Value from which to plot AVAL values

input_plot_adam
A character string containing the ADaM dataset name

input_overlay_events
A character vector containing the names of patient events to plot

vline_dat
The vline data frame that contains x-intercept values for the corrresponding events selected to be overlain on the plot

vv_dy_name
TA character vector containing the name of the visit variable(s)

Value

If graph_output is plotly, then a plotly object, else if ggplot then a ggplot2 object

See also