Gather Rules For Help UI

Gather Rules into a shiny tagList to be included in the UI for "help" purposes, conditionally containing different required and/or recommended rule sets (if they exist).

gather_rules(
  input,
  output,
  session,
  all_df_rules = list(error = c(""), warn = c("")),
  expl_rules = list(list(error = c(""), warn = c(""))),
  df_incl_rules = list(list(error = c(""), warn = c("")))
)

Arguments

input, output, session
Internal parameters for shiny.

all_df_rules
A double-nested list: inner list is named list of variables names that should result in error or warn if variables do not exist or are missing for ANY data frame. Outter list is unnamed.

expl_rules
A double-nested list: outter list is a named list dataframes. Inner list (for each data frame) contains named lists of variables names that should result in error or warn if variables do not exist or are missing for the specified parent data frame

df_incl_rules
A double-nested list: outter list is a named list variable names used to identify a particular class of data frame. For example, the variable name PARAMCD would id BDS class data frames, and the inner list would contain a named lists of variables names that should result in error or warn if variables do not exist or are missing for the implied parent data frame.

Value

An shiny tagList

See also