Table Generator Cicerone R6 Object

This object is used within the table generator module to add help text to the various fields using a help buttom

tg_guide

Format

An object of class Cicerone (inherits from R6) of length 13.

See also